backend/requirements.txt
asgiref==3.11.0
astroid==4.0.2
cfgv==3.5.0
dill==0.4.0
distlib==0.4.0
django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
filelock==3.20.1
identify==2.6.15
isort==7.0.0
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.5.1
pre-commit==4.5.1
pyjwt==2.10.1
pylint==4.0.4
pyyaml==6.0.3
sqlparse==0.5.4
tomlkit==0.13.3
virtualenv==20.35.4
